refactor: Enable exercise in courses. Rename files.
