+++ apps/web/src/serlo-editor-integration/serlo-plugin-wrappers/exercise-serlo-static-renderer.tsx
@@ -9,7 +9,7 @@ import { useContext, useEffect, useState } from 'react'
 import { useAuthentication } from '@/auth/use-authentication'
 import { ExerciseLicenseNotice } from '@/components/content/license/exercise-license-notice'
 import type { MoreAuthorToolsProps } from '@/components/user-tools/foldout-author-menus/more-author-tools'
-import { ExerciseContext } from '@/contexts/exercise-ids-context'
+import { ExerciseContext } from '@/contexts/exercise-context'
 import { RevisionViewContext } from '@/contexts/revision-view-context'
 import { ExerciseInlineType } from '@/data-types'
