+++ apps/web/src/serlo-editor-integration/serlo-plugin-wrappers/h5p-serlo-static.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/router'
 import { useContext, useEffect } from 'react'
 
 import { useAB } from '@/contexts/ab'
-import { ExerciseContext } from '@/contexts/exercise-ids-context'
+import { ExerciseContext } from '@/contexts/exercise-context'
 import { useEntityData } from '@/contexts/uuids-context'
 import { exerciseSubmission } from '@/helper/exercise-submission'
 import { useCreateExerciseSubmissionMutation } from '@/mutations/use-experiment-create-exercise-submission-mutation'
