+++ apps/web/src/serlo-editor-integration/serlo-plugin-wrappers/sc-mc-serlo-static-renderer.tsx
@@ -6,7 +6,7 @@ import { useContext } from 'react'
 
 import { isPrintMode } from '@/components/print-mode'
 import { useAB } from '@/contexts/ab'
-import { ExerciseContext } from '@/contexts/exercise-ids-context'
+import { ExerciseContext } from '@/contexts/exercise-context'
 import { useInstanceData } from '@/contexts/instance-context'
 import { RevisionViewContext } from '@/contexts/revision-view-context'
 import { useEntityData } from '@/contexts/uuids-context'
